requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pydantic-settings==2.0.3
aiofiles==23.2.1
pillow==10.1.0
httpx==0.25.2

# OSS和S3支持
oss2==2.18.4
boto3==1.34.0

# Redis支持（可选）
redis==5.0.1